constructor: break down libp2p logic
License: MIT
